+++ src/components/MainNav/AppOrderProvider.js
@@ -103,7 +103,7 @@ function getAllowedApps(appModules, stripes, pathname, lastVisited, formatMessag
       route: SETTINGS_ROUTE
     });
   }
-  return apps;
+  return apps.toSorted((a, b) => a.displayName.localeCompare(b.displayName));
 }
 
 /**