+++ src/components/PreLoginLanding/PreLoginLanding.js
@@ -1,17 +1,19 @@
 import React, { useRef } from 'react';
-import { Button, Select, Col, Row } from '@folio/stripes-components';
 import { useIntl } from 'react-intl';
 import PropTypes from 'prop-types';
-import { OrganizationLogo } from '../index';
-import styles from './index.css';
+
+import { Button, Select, Col, Row } from '@folio/stripes-components';
+import OrganizationLogo from '../OrganizationLogo';
 import { useStripes } from '../../StripesContext';
+import { getOIDCRedirectUri } from '../../loginServices';
+import styles from './index.css';
 
 function PreLoginLanding({ onSelectTenant }) {
   const intl = useIntl();
   const { okapi, config: { tenantOptions = {} } } = useStripes();
 
-  const redirectUri = `${window.location.protocol}//${window.location.host}/oidc-landing`;
-  const options = Object.keys(tenantOptions).map(tenantName => ({ value: tenantName, label: tenantName }));
+  const redirectUri = getOIDCRedirectUri(okapi.tenant, okapi.clientId);
+  const options = Object.values(tenantOptions).map(i => ({ value: i.name, label: i.displayName ?? i.name }));
 
   const getLoginUrl = () => {
     if (!okapi.tenant) return '';