+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 # Change history for stripes-core
 
 ## 11.1.0 IN PROGRESS
+
 * CSS Support for printing of results lists content. Refs STCOR-956.
 * Improve useModuleInfo hook. Refs STCOR-955.
+* Provide `useQueryLimit()` hook. Refs STCOR-616, STCOR-617.
+* Show user-friendly labels on ECS pre-login screen. Refs STCOR-899.
+* Sort app links in the main navigation by their `displayName` by default. Refs STCOR-964.
+* Change default Idle Session Timeout to `4h` from `1h`. Refs STCOR-962.
+* Prune unused `./utils/` scripts, deps. Refs STCOR-959.
+* Don't show IST modal if `flsTimeRemaining` less than config.rtr.idleModalTTL. STCOR-884.
+* Allow customizing request timeout in `useOkapiKy`. STCOR-967.
+* Enhance locale management to support user and tenant locale configurations from both mod-configuration and mod-settings. STCOR-968.
+* When a session-restart request fails, trap errors and purge both redux and browser-storage. Refs STCOR-970.
+* For OIDC login - store `tenant` and `clientId` in `redirect_uri` url parameter, instead of localStorage. Refs STCOR-972.
+* Disclose personal data disclosure (i.e. user data cached in storage). Refs STCOR-973.
+* Removed code that switches locale, numbering system etc to tenant settings when user doesn't have a locale preference. Refs STCOR-981.
 
 ## [11.0.0](https://github.com/folio-org/stripes-core/tree/v11.0.0) (2025-02-24)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v10.2.0...v11.0.0)